+++ CHANGELOG.md
@@ -1,4 +1,8 @@
+## 0.3.0
+
+* Add `SystemChineseFont.textTheme` for `ThemeData`
+  
 ## 0.2.0
 
-* Support for default Chinese font family on iOS, macOS, Xiaomi
-* Support for dynamically load fonts from url, file and asset
+* Support for system Chinese font on Apple, Xiaomi, and Windows
+* Support for dynamically load fonts from url, file, and asset